restructure(grid): moved coordinate systems methods to separate utility class Geometry
