+++ Assets/Scripts/Geometry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49ad9a1a1144e474d8cdcc823ce93ecd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
