+++ Assets/Scripts/RoomBuilder.cs
@@ -13,7 +13,7 @@ public class RoomBuilder : MonoBehaviour
     public void BuildRoom(Vector2Int gridIntPosition, GameObject roomPrefab, Settings settings)
     {
         GameObject newRoom = Instantiate(roomPrefab);
-        newRoom.transform.position = GridToWorldPosition(gridIntPosition + Vector2Int.one, settings);
+        newRoom.transform.position = Geometry.GridToWorldPosition(gridIntPosition + Vector2Int.one, settings);
     }
 
     private void Update()
@@ -30,7 +30,7 @@ public class RoomBuilder : MonoBehaviour
         Settings settings = settingsManager.Settings;
         
         Vector2 pointerWorldPosition = camera.ScreenToWorldPoint(Input.mousePosition);
-        Vector2 pointerGridPosition = WorldToGridPosition(pointerWorldPosition, settings);
+        Vector2 pointerGridPosition = Geometry.WorldToGridPosition(pointerWorldPosition, settings);
         Vector2Int pointerGridIntPosition = Vector2Int.FloorToInt(pointerGridPosition);
         
         bool inputBuildRoom = false;
@@ -55,7 +55,7 @@ public class RoomBuilder : MonoBehaviour
         if (_dragAndDropRoomPreview)
         {
             _dragAndDropRoomPreview.position =
-                GridToWorldPosition(Vector2Int.FloorToInt(pointerGridPosition) + Vector2Int.one, settings);
+                Geometry.GridToWorldPosition(Vector2Int.FloorToInt(pointerGridPosition) + Vector2Int.one, settings);
         }
 
         if (inputBuildRoom)
@@ -67,40 +67,5 @@ public class RoomBuilder : MonoBehaviour
     private void Start()
     {
         _dragAndDropRoomPreview = Instantiate(DebugGrayRoomPrefab).transform;
-    }
-
-    public static Vector2 WorldToGridPosition(Vector3 worldPosition, Settings settings)
-    {
-        Vector2 gridXInWorldSpace = settings.GridSpaceXInWorldSpace;
-        Vector2 gridYInWorldSpace = settings.GridSpaceYInWorldSpace;
-        // Grid Space Matrix:
-        // { gridXInWorldSpace.x, gridYInWorldSpace.x }
-        // { gridXInWorldSpace.y, gridYInWorldSpace.y }
-        // Getting inverse Matrix, because:
-        // gridSpacePosition = inverseGridSpaceMatrix * worldPosition
-        // I will code matrix as two 2d vectors: 'x' and 'y'.
-        // So 'x' is the first column of matrix and 'y' is the second.
-        float gridSpaceDeterminant = gridXInWorldSpace.x * gridYInWorldSpace.y - gridXInWorldSpace.y * gridYInWorldSpace.x;
-        Vector2 adjugateX = new(gridYInWorldSpace.y, -gridXInWorldSpace.y);
-        Vector2 adjugateY = new(-gridYInWorldSpace.x, gridXInWorldSpace.x);
-        Vector2 inverseGridSpaceMatrixX = adjugateX / gridSpaceDeterminant;
-        Vector2 inverseGridSpaceMatrixY = adjugateY / gridSpaceDeterminant;
-        float xGridSpace = inverseGridSpaceMatrixX.x * worldPosition.x + inverseGridSpaceMatrixY.x * worldPosition.y;
-        float yGridSpace = inverseGridSpaceMatrixX.y * worldPosition.x + inverseGridSpaceMatrixY.y * worldPosition.y;
-        return new Vector2(xGridSpace, yGridSpace);
-    }
-
-    public static Vector2 GridToWorldPosition(Vector2 gridPosition, Settings settings)
-    {
-        Vector2 gridXInWorldSpace = settings.GridSpaceXInWorldSpace;
-        Vector2 gridYInWorldSpace = settings.GridSpaceYInWorldSpace;
-        // Grid Space Matrix:
-        // { gridXInWorldSpace.x, gridYInWorldSpace.x }
-        // { gridXInWorldSpace.y, gridYInWorldSpace.y }
-        // Formula:
-        // worldCoordinates = transformMatrix * gridCoordinates
-        float x = gridXInWorldSpace.x * gridPosition.x + gridYInWorldSpace.x * gridPosition.y;
-        float y = gridXInWorldSpace.y * gridPosition.x + gridYInWorldSpace.y * gridPosition.y;
-        return new Vector2(x, y);
     }
 }
